proyecto terminado

+++ app/Http/Controllers/VeedoresController.php
@@ -15,6 +15,14 @@ use function Spatie\Ignition\ErrorPage\report;
 
 class VeedoresController extends Controller
 {
+    public function indexview(Request $request)
+    {
+        return view('layouts.dashboard.index');
+    }
+    public function detail(Request $request)
+    {
+        return view('layouts.dashboard.index');
+    }
     /**
      * Display a listing of the resource.
      *
@@ -22,7 +30,8 @@ class VeedoresController extends Controller
      */
     public function index()
     {
-        $veedores = DB::select('CALL view_veedores()');
+        $user = \Auth::guard('api')->user();
+        $veedores = DB::select('CALL view_veedores(?)', array($user->id));
 
         if ($veedores) {
             return response()->json(['status' => 'success', 'veedores' => $veedores]);
@@ -51,7 +60,7 @@ class VeedoresController extends Controller
     {
 
         if (!is_array($request->all())) {
-            return ['error' => 'request must be an array'];
+            return ['status' => 'error','message'=>'request must be an array'];
         }
 
         $rules = [
@@ -74,7 +83,7 @@ class VeedoresController extends Controller
             if ($validator->fails()) {
                 return [
                     'status' => 'Error',
-                    'errors' => $validator->errors()->all()
+                    'message' => $validator->errors()->all()
                 ];
             }
 
@@ -86,22 +95,42 @@ class VeedoresController extends Controller
             $filename_f = $veedor->imagen_frontal->getClientOriginalName();
             $filename_r = $veedor->imagen_reverso->getClientOriginalName();
 
-            $save_path = storage_path('app/public') . '/veedores/dni/' . $veedor->dni . '/upload/imagenes/';
-            $public_path_f = '/veedores/dni/' . $veedor->dni . '/upload/imagenes/' . $filename_f;
-            $public_path_r = '/veedores/dni/' . $veedor->dni . '/upload/imagenes/' . $filename_r;
+            //$save_path = storage_path('app/public') . '/veedores/dni/' . $veedor->dni . '/upload/imagenes/';
+            $save_path =  '/veedores/dni/' . $veedor->dni . '/uploads/imagenes/';
+            // $public_path_f = '/veedores/dni/' . $veedor->dni . '/upload/imagenes/' . $filename_f;
+            // $public_path_r = '/veedores/dni/' . $veedor->dni . '/upload/imagenes/' . $filename_r;
+            $public_path_f = $save_path . $filename_f;
+            $public_path_r = $save_path . $filename_r;
 
-            File::makeDirectory($save_path, $mode = 0755, true, true);
-            Image::make($veedor->imagen_frontal)->save($save_path . $filename_f);
-            Image::make($veedor->imagen_reverso)->save($save_path . $filename_r);
-
+            // File::makeDirectory($save_path, $mode = 0755, true, true);
+            // Image::make($veedor->imagen_frontal)->save($save_path . $filename_f);
+            // Image::make($veedor->imagen_reverso)->save($save_path . $filename_r);
+            $path = Storage::putFileAs(
+                'public'.$save_path,
+                $veedor->imagen_frontal,
+                $filename_f
+            );
+            if (!$path) {
+                \DB::rollback();
+                return response()->json(array("status" => "error",'message'=>'Hubo un error al actualizar'));
+            }
+            $path = Storage::putFileAs(
+                'public/'.$save_path,
+                $veedor->imagen_reverso,
+                $filename_r
+            );
+            if (!$path) {
+                \DB::rollback();
+                return response()->json(array("status" => "error",'message'=>'Hubo un error al actualizar'));
+            }
             $veedor->imagen_frontal = $public_path_f;
             $veedor->imagen_reverso = $public_path_r;
 
             $veedor->save();
-            return ['status' => 'success'];
-        } catch (\Exception $e) {
+            return ['status' => 'success','message'=>'veedor registrado'];
+        } catch (\Throwable $e) {
             \Log::info('Error creating user: ' . $e);
-            return \Response::json(['status' => 'Error'], 500);
+            return \Response::json(['status' => 'error','message'=>$e->getMessage()], 500);
         }
     }
 
@@ -114,11 +143,10 @@ class VeedoresController extends Controller
     public function show(Request $request)
     {
         $veedor = DB::select('CALL sp_view_veedores(?)', array($request->id));
-
         if ($veedor) {
-            return response()->json(['status' => 'sucess', 'veedor' => $veedor]);
+            return response()->json(['status' => 'sucess', 'veedor' => $veedor[0]]);
         } else {
-            return response()->json(['status' => 'Failed, El usuario no existe']);
+            return response()->json(['status' => 'error','message'=>'Failed, El usuario no existe']);
         }
     }
 
@@ -141,45 +169,60 @@ class VeedoresController extends Controller
     {
         try {
             if ($veedor) {
-                if ($request->hasFile('imagen_frontal') || $request->hasFile('imagen_reverso')) {
-                    $filename_f = $veedor->imagen_frontal;
-                    $filename_r = $veedor->imagen_reverso;
-
-                    if ($filename_f) {
-                        Storage::disk('public')->delete($filename_f);
-                    } else if ($filename_r) {
-                        Storage::disk('public')->delete($filename_r);
-                    } else if ($filename_f && $filename_r) {
-                        Storage::disk('public')->delete($filename_f);
-                        Storage::disk('public')->delete($filename_r);
+                if($request->hasFile('imagen_frontal') || $request->hasFile('imagen_reverso')){
+                    if ($request->hasFile('imagen_frontal')) {
+                        $filename_f = $veedor->imagen_frontal;
+    
+                        if ($filename_f) {
+                            Storage::disk('public')->delete($filename_f);
+                        }
+    
+                        $veedor->fill($request->validated());
+                        $veedor->imagen_frontal = $request->file('imagen_frontal');
+                        $filename_f = $veedor->imagen_frontal->getClientOriginalName();
+                        $save_path = '/veedores/dni/' . $veedor->dni . '/uploads/imagenes/';
+                        $public_path_f = $save_path . $filename_f;
+                        $path = Storage::putFileAs(
+                            'public'.$save_path,
+                            $veedor->imagen_frontal,
+                            $filename_f
+                        );
+                        if (!$path) {
+                            \DB::rollback();
+                            return response()->json(array("status" => "error",'message'=>'Hubo un error al actualizar'));
+                        }
+                        // File::makeDirectory($save_path, $mode = 0755, true, true);
+                        // Image::make($veedor->imagen_frontal)->save($save_path . $filename_f);
+                        // Image::make($veedor->imagen_reverso)->save($save_path . $filename_r);
+                        $veedor->imagen_frontal = $public_path_f;
                     }
-
-
-                    $veedor->fill($request->validated());
-
-                    $veedor->imagen_frontal = $request->file('imagen_frontal');
-                    $veedor->imagen_reverso = $request->file('imagen_reverso');
-
-                    $filename_f = $veedor->imagen_frontal->getClientOriginalName();
-                    $filename_r = $veedor->imagen_reverso->getClientOriginalName();
-
-                    $save_path = storage_path('app/public') . '/veedores/dni/' . $veedor->dni . '/uploads/imagenes/';
-                    $public_path_f = '/veedores/dni/' . $veedor->dni . '/uploads/imagenes/' . $filename_f;
-                    $public_path_r = '/veedores/dni/' . $veedor->dni . '/uploads/imagenes/' . $filename_r;
-
-                    File::makeDirectory($save_path, $mode = 0755, true, true);
-                    Image::make($veedor->imagen_frontal)->save($save_path . $filename_f);
-                    Image::make($veedor->imagen_reverso)->save($save_path . $filename_r);
-
-                    $veedor->imagen_frontal = $public_path_f;
-                    $veedor->imagen_reverso = $public_path_r;
-
+                    if($request->hasFile('imagen_reverso')){
+                        $filename_r = $veedor->imagen_reverso;
+                        if ($filename_r) {
+                            Storage::disk('public')->delete($filename_r);
+                        }
+                        $veedor->fill($request->validated());
+                        $veedor->imagen_reverso = $request->file('imagen_reverso');
+                        $filename_r = $veedor->imagen_reverso->getClientOriginalName();
+                        $save_path = '/veedores/dni/' . $veedor->dni . '/uploads/imagenes/';
+                        $public_path_r = $save_path . $filename_r;
+                        $path = Storage::putFileAs(
+                            'public'.$save_path,
+                            $veedor->imagen_reverso,
+                            $filename_r
+                        );
+                        if (!$path) {
+                            \DB::rollback();
+                            return response()->json(array("status" => "error",'message'=>'Hubo un error al actualizar'));
+                        }
+                        $veedor->imagen_reverso = $public_path_r;
+                    }
                     $res = $veedor->save();
-                } else {
+                }else {
                     $res = $veedor->update(array_filter($request->validated()));
                 }
             } else {
-                return response()->json(['status' => 'Usuario no encontrado']);
+                return response()->json(['status' => 'error','message'=>'Usuario no encontrado']);
             }
         } catch (\Throwable $th) {
             DB::rollback();
@@ -187,7 +230,7 @@ class VeedoresController extends Controller
         }
 
         if ($res) {
-            return response()->json(['status' => 'Veedor actualizado']);
+            return response()->json(['status' => 'success','message'=>'Veedor actualizado']);
         }
     }
 
@@ -205,13 +248,13 @@ class VeedoresController extends Controller
             if ($veedor->imagen_frontal && $veedor->imagen_reverso) {
                 File::deleteDirectory(storage_path('app/public') . '/users/dni/' . $veedor->dni);
                 $veedor->delete();
-                return response()->json(['status' => 'Eliminado']);
+                return response()->json(['status' => 'success','message'=>'Eliminado']);
             } else {
                 $veedor->delete();
-                return response()->json(['status' => 'Eliminado']);
+                return response()->json(['status' => 'success','message'=>'Eliminado']);
             }
         } else {
-            return response()->json(['status' => 'El veedor no ha sido encontrado']);
+            return response()->json(['status' => 'error','message'=>'El veedor no ha sido encontrado']);
         }
     }
 }
